+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For details about compatibility between different releases, see the **Commitment
 ## [Unreleased]
 
 ### Added
+- CLI support for federated authentication provider deletion (`ttn-lw-stack is-db delete-auth-provider`).
 
 ### Changed